requirements.txt
opencv-python
torch
torchvision
numpy
scikit-learn
matplotlib
Pillow
torchsummary
polarTransform
imageio
tqdm
linformer
vit_pytorch
tkinter
